requirements.txt
Flask
gunicorn
flask-cors
# llama-index
# llama-index-llms-groq
# llama-index-embeddings-huggingface
# llama-index-embeddings-openai
python-dotenv
numpy==1.23.5
pandas==2.3.1
scikit-learn==1.7.1
tensorflow==2.12.0
keras==2.12.0
ultralytics==8.3.170
torch==2.7.1
torchvision==0.22.1
Pillow==11.3.0
